Be clear about what each procedure needs
Up to now I have glommed methods onto the ReleaseContext when I
factored out a procedure. But not many of them actually need it, so I
have separated things out into

 - helpers for the release job handler that need to update the job
   while they work (left in releaser.go);
 - procedures that just do a calculation on some data;
 - procedures that operate on a `ReleaseContext`'s working
   directory, and therefore belong on `ReleaseContext` (in context.go);
 - procedures that need an `instance.Instance` or closely related type
   to get some data or apply some changes (in instance.go).
