+++ release/context_test.go
@@ -8,7 +8,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/weaveworks/flux"
 	"github.com/weaveworks/flux/git"
 	"github.com/weaveworks/flux/instance"
 	"github.com/weaveworks/flux/platform/kubernetes/testdata"
@@ -40,34 +39,6 @@ func TestCloneCommitAndPush(t *testing.T) {
 	err = ctx.CommitAndPush("Removed file")
 	if err != nil {
 		t.Fatal(err)
-	}
-}
-
-func TestLockedServices(t *testing.T) {
-	conf := instance.Config{
-		Services: map[flux.ServiceID]instance.ServiceConfig{
-			flux.ServiceID("service1"): instance.ServiceConfig{
-				Locked: true,
-			},
-			flux.ServiceID("service2"): instance.ServiceConfig{
-				Locked:    true,
-				Automated: true,
-			},
-			flux.ServiceID("service3"): instance.ServiceConfig{
-				Automated: true,
-			},
-		},
-	}
-
-	locked := LockedServices(conf)
-	if !locked.Contains(flux.ServiceID("service1")) {
-		t.Error("service1 locked in config but not reported as locked")
-	}
-	if !locked.Contains(flux.ServiceID("service2")) {
-		t.Error("service2 locked in config but not reported as locked")
-	}
-	if locked.Contains(flux.ServiceID("service3")) {
-		t.Error("service3 not locked but reported as locked")
 	}
 }
